Break out the comments into another file
